transferring project from codepen still (check desc)
this css was *rewritten* by chatgpt, but it rewrote *my* code. i was having issues after adding some javascript into the project, and after over an hour of trying to bugfix that issue, i caved in and used chatgpt's rendition which had worked. im sorry for stooping that low, but i was on my LAST limb 😭